Added support for sending the lang parameter.
This adds a couple more optional methods you can use, the main one is:

   [_weatherAPI setLangWithPreferedLanguage];

it sets the language code send to the api according to the language on
the phone.

you can also manually set it with the method

    - (void) setLang:(NSString *) lang

setting it to nil will make the api send no lang parameter to the api.